+++ sequencer/src/api.rs
@@ -695,6 +695,7 @@ pub mod test_helpers {
     };
     use hotshot::types::{Event, EventType};
     use hotshot_contract_adapter::sol_types::{LightClientStateSol, StakeTableStateSol};
+    use hotshot_state_prover::service::legacy_light_client_genesis_from_stake_table;
     use hotshot_types::{
         event::LeafInfo,
         traits::{metrics::NoMetrics, node_implementation::ConsensusTime},
@@ -722,38 +723,6 @@ pub mod test_helpers {
     };
 
     pub const STAKE_TABLE_CAPACITY_FOR_TEST: u64 = 10;
-
-    use hotshot_contract_adapter::field_to_u256;
-    use hotshot_stake_table::vec_based::StakeTable;
-    use hotshot_types::{
-        light_client::{one_honest_threshold, CircuitField, StateVerKey},
-        signature_key::BLSPubKey,
-        traits::stake_table::{SnapshotVersion, StakeTableScheme},
-    };
-
-    #[inline]
-    fn light_client_genesis_from_stake_table_for_test(
-        st: StakeTable<BLSPubKey, StateVerKey, CircuitField>,
-    ) -> anyhow::Result<(LightClientStateSol, StakeTableStateSol)> {
-        let (bls_comm, schnorr_comm, stake_comm) = st
-            .commitment(SnapshotVersion::LastEpochStart)
-            .expect("Commitment computation shouldn't fail.");
-        let threshold = one_honest_threshold(st.total_stake(SnapshotVersion::LastEpochStart)?);
-
-        Ok((
-            LightClientStateSol {
-                viewNum: 0,
-                blockHeight: 0,
-                blockCommRoot: U256::from(0u32),
-            },
-            StakeTableStateSol {
-                blsKeyComm: field_to_u256(bls_comm),
-                schnorrKeyComm: field_to_u256(schnorr_comm),
-                amountComm: field_to_u256(stake_comm),
-                threshold,
-            },
-        ))
-    }
 
     pub struct TestNetwork<P: PersistenceOptions, const NUM_NODES: usize, V: Versions> {
         pub server: SequencerContext<network::Memory, P::Persistence, V>,
@@ -996,7 +965,7 @@ pub mod test_helpers {
 
         pub fn light_client_genesis(&self) -> (LightClientStateSol, StakeTableStateSol) {
             let st = self.cfg.stake_table();
-            light_client_genesis_from_stake_table_for_test(st).unwrap()
+            legacy_light_client_genesis_from_stake_table(st).unwrap()
         }
 
         pub async fn stop_consensus(&mut self) {
